chore(query): support spill by max memory

+++ src/query/service/src/pipelines/processors/transforms/aggregator/transform_aggregate_partial.rs
@@ -73,7 +73,7 @@ impl TryFrom<Arc<QueryContext>> for AggregateSettings {
         let settings = ctx.get_settings();
         let convert_threshold = settings.get_group_by_two_level_threshold()? as usize;
         let value = settings.get_spilling_bytes_threshold_per_proc()?;
-        let max_memory_usage = settings.get_max_memory_usage()?;
+        let max_memory_usage = settings.get_max_memory_usage()? as f64;
 
         Ok(AggregateSettings {
             convert_threshold,
